Add CLI options to alphagen and move into main workspace
