+++ alphagen/Cargo.toml
@@ -7,7 +7,6 @@ description = "Used to generate greyscale mask images based on the alpha channel
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = { version = "3", features = ["derive"] }
 image = "*"
 rayon = "*"
-
-[workspace]
